init: start project React Router
