+++ API/package.json
@@ -47,6 +47,7 @@
     "mini-css-extract-plugin": "^2.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0",
     "styled-components": "^5.3.0"
   }
 }
